requirements.txt
transformers==4.42.2
sentencepiece
tiktoken
protobuf
hazm
pandas
scikit-learn
arabic-reshaper
python-bidi
